% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\alias{Data.default}
\alias{Data.outRanger}
\title{Data of outRanger}
\usage{
Data(object, ...)

\method{Data}{default}(object, ...)

\method{Data}{outRanger}(object, ...)
}
\arguments{
\item{object}{An object of class \code{outRanger}.}

\item{...}{Arguments passed from or to other methods.}
}
\value{
A \code{data.frame}.
}
\description{
Extracts data with replaced outliers from object of class \code{outRanger}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method not implemented yet.

\item \code{outRanger}: Extract data from outRanger object.
}}

\examples{
x <- outRanger(iris)
head(Data(x))
}
