% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.outRanger}
\alias{is.outRanger}
\title{Check function}
\usage{
is.outRanger(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
A logical vector of length one.
}
\description{
Checks if an object inherits class \code{outRanger}.
}
\examples{
a <- outRanger(iris)
is.outRanger(a)
is.outRanger("a")
}
