% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\alias{outliers.default}
\alias{outliers.outRanger}
\title{outliers of outRanger}
\usage{
outliers(object, ...)

\method{outliers}{default}(object, ...)

\method{outliers}{outRanger}(object, ...)
}
\arguments{
\item{object}{An object of class \code{outRanger}.}

\item{...}{Arguments passed from or to other methods.}
}
\value{
A \code{data.frame}. Each row describes an outlier. The columns are as follows:
\itemize{
  \item \code{row}, \code{col} Row and column in original data with outlier.
  \item \code{observed} Observed value.
  \item \code{predicted} Predicted value.
  \item \code{rmse} Scaling factor used to normalize the difference between observed and predicted.
  \item \code{score} Outlier score defined as (observed-predicted)/rmse.
  \item \code{replacement} Value used to replace observed value.
}
}
\description{
Extracts outliers from object of class \code{outRanger}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method not implemented yet.

\item \code{outRanger}: Extract outliers from outRanger object.
}}

\examples{
x <- outRanger(iris)
outliers(x)
}
