% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outRanger.R
\name{outRanger}
\alias{outRanger}
\title{Multivariate Outlier Detection and Replacement by Random Forest Predictions}
\usage{
outRanger(
  data,
  formula = . ~ .,
  replace = c("pmm", "predictions", "NA", "no"),
  pmm.k = 3,
  threshold = 3,
  max_n_outliers = Inf,
  max_prop_outliers = 1,
  min.node.size = 40,
  impute_multivariate = TRUE,
  impute_multivariate_control = list(pmm.k = 3, num.trees = 20, maxiter = 2L),
  seed = NULL,
  verbose = 1,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} to be assessed for numeric outliers.}

\item{formula}{A two-sided formula specifying variables to be checked (left hand side) and variables used to check (right hand side). Defaults to . ~ ., i.e. use all variables to check all (numeric) variables.}

\item{replace}{Should outliers be replaced by predicting mean matching on the OOB predictions ("pmm", the default), by OOB predictions ("predictions"), by \code{NA} ("NA"). Use "no" to keep outliers as they are.}

\item{pmm.k}{For \code{replace = "pmm"}, how many nearest prediction neighbours (without outliers) be considered to sample observed values from?}

\item{threshold}{Threshold above which an outlier score is considered an outlier. The default is 3.}

\item{max_n_outliers}{Maximal number of outliers to identify. Will be used in combination with \code{threshold} and \code{max_prop_outliers}.}

\item{max_prop_outliers}{Maximal relative count of outliers. Will be used in combination with \code{threshold} and \code{max_n_outliers}.}

\item{min.node.size}{Minimal node size of the random forests. With 40, the value is relatively high. This reduces the impact of outliers.}

\item{impute_multivariate}{If \code{TRUE} (default), missing values are imputed by \code{missRanger::missRanger}. Otherwise, by univariate sampling.}

\item{impute_multivariate_control}{Parameters passed to \code{missRanger::missRanger} if data contains missing values.}

\item{seed}{Integer random seed.}

\item{verbose}{Controls how much outliers is printed to screen. 0 to print nothing, 1 prints information.}

\item{...}{Arguments passed to \code{ranger}. If the data set is large, use less trees (e.g. \code{num.trees = 20}) and/or a low value of \code{mtry}.
The following arguments are e.g. incompatible with \code{ranger}: \code{write.forest}, \code{probability}, \code{dependent.variable.name}, and \code{classification}.}
}
\value{
An object of type \code{outRanger} (and a list). The data set with replaced values can be extracted by calling \code{Data}. A description of all outliers detected can be found by calling \code{outliers}.
}
\description{
This function provides a random forest based implementation of the method described in Chapter 7.1.2 ("Regression Model Based Anomaly detection") of [1]. Each numeric variable to be checked for outliers is regressed onto all other variables using a random forest. If the scaled absolute difference between observed value and out-of-bag prediction is larger than some predefined threshold (default is 3), then a value is considered an outlier, see Details below. After identification of outliers, they can be replaced e.g. by predictive mean matching from the non-outliers. Since the random forest algorithm "ranger" [2] does not allow for missing values, any missing value is first being imputed by chained random forests. The method can be viewed as a multivariate extension of a basic univariate outlier detection method where a value is considered an outlier if it is more than e.g. three times the standard deviation away from its mean. In the multivariate case, instead of comparing a value with the overall mean, rather the difference to the conditional mean is considered. The 'outRanger' function estimates this conditional mean by a random forest.
}
\details{
The outlier score of the i-th value x_ij of the j-th variable is defined as s_ij = (x_ij - pred_ij) / rmse_j, where pred_ij is the corresponding out-of-bag prediction of the j-th random forest and rmse_j its RMSE. If |s_ij| > L with threshold L, then x_ij is considered an outlier.
For large data sets, just by chance, many values can surpass the default threshold of 3. To reduce the number of outliers, the threshold can be increased. Alternatively, the number of outliers can be limited by the two arguments \code{max_n_outliers} and \code{max_prop_outliers}. E.g. if at most ten outliers are to be identified, set \code{max_n_outliers = 10}.
}
\examples{
head(irisWithOut <- generateOutliers(iris, seed = 345))
(out <- outRanger(irisWithOut))
outliers(out)
head(Data(out))
plot(out)
plot(out, what = "scores")
}
\references{
[1] Chandola V., Banerjee A., and Kumar V. (2009). Anomaly detection: A survey. ACM Comput. Surv. 41, 3, Article 15 <dx.doi.org/10.1145/1541880.1541882>.
[2] Wright, M. N. & Ziegler, A. (2016). ranger: A Fast Implementation of Random Forests for High Dimensional Data in C++ and R. Journal of Statistical Software, in press. <arxiv.org/abs/1508.04409>.
}
\seealso{
\code{\link{outliers}}, \code{\link{Data}}, \code{\link{plot.outRanger}}, \code{\link{summary.outRanger}}.
}
