% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.outRanger}
\alias{print.outRanger}
\title{Prints outRanger}
\usage{
\method{print}{outRanger}(x, ...)
}
\arguments{
\item{x}{A on object of class \code{outRanger}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
Invisibly, the input is returned.
}
\description{
Print method for an object of class \code{outRanger}.
}
\examples{
x <- outRanger(iris)
x
}
