% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.outRanger}
\alias{summary.outRanger}
\title{Summarizes outRanger}
\usage{
\method{summary}{outRanger}(object, ...)
}
\arguments{
\item{object}{A on object of class \code{outRanger}.}

\item{...}{Further arguments passed from other methods.}
}
\value{
A list of summary statistics.
}
\description{
Summary method for an object of class \code{outRanger}. Besides the number of outliers per variables, it shows univariate distributions of the values considered as outliers.
}
\examples{
object <- outRanger(iris, seed = 34, verbose = 0)
summary(object)
}
